Fix 64-bit issues where &int was given instead of &gsize. Fixes crash on AMD64 (and possibly Sparc?).
git-svn-id: https://xchat.svn.sourceforge.net/svnroot/xchat@475 893a96be-7f27-4fdf-9d1e-6aeec9d3cce1

+++ src/common/server.c
@@ -289,8 +289,8 @@ server_inline (server *serv, char *line, int len)
 		{
 			char *conv_line; /* holds a copy of the original string */
 			int conv_len; /* tells g_convert how much of line to convert */
-			int utf_len;
-			int read_len;
+			gsize utf_len;
+			gsize read_len;
 			GError *err;
 			gboolean retry;
